Rename `reporting_lag_period` -> `reporting_delay_period`
